Clarify what a component and element are (#1719)
* Clarify what a component and element are

We weren't being explicit about this. Studying more React on my own and with others, I think it makes sense to add a little clarification that `Cat` is the component and `<Text>` is the React element—that represents Views. Would've loved to put this in the prior section on Native Components and Core Components, but honestly, we really needed to explain that RN isn't a compiler and what Views are in relation to Native Components. Hopefully this little bit helps drive this home even more :)

* Merging it a little better.