docs: add fragment to native-components-android.md (#2599)
* docs: add fragment to native-components-android.md

As a Javascript / React-Native developer I had a hard time figuring out how to integrate a custom native SDK to React Native with the help of Android Fragments. Please correct anything that you wish. This information took me 3 days to figure out, especially the part with `setupLayout` and `manuallyLayoutChildren` so I wish an example like this was available in the official documentation. I was expecting the code without `setupLayout` and `manuallyLayoutChildren` but i had no luck. If somebody has any better solution in your team, or any explanation, it would also be really nice to state in the documentation

* chore: fix linting errors

* chore: fix alex errors

* docs: update native-components-android.md

* docs: update native-components-android.md

- add missing imports
- remove extra spaces
- remove unnecessary double cast
- remove unnecessary comment