Start TypeScriptify-ing the examples (#3450)
* Start TypeScriptify-ing the examples

This change starts migrating the examples to TypeScript. Every example is now linted as both tsx and jsx, unless an extension is specified. All of the existing ESLint warnings, minus those for inline styles which I have disabled for now, are addressed, so we can now flag new lint warnings when examples are added. A script `tsc-examples` is added, in addition to ESLint, but I haven't folded it into `lint:examples` yet because not everything has been fixed, and doing so will take multiple PRs.

100/171 examples are clean for both ESLint and TypeScript in strict mode, with an additional in this change opted out of typechecking (Clipboard) since it has been removed and is no longer part of typings. Most examples are already valid as both JSX and strict TSX, but some are separated into tabs (where TypeScript is shown first and by deafult), require changes so that TypeScript can infer the types (while still being valid JS), or require fixup (e.g. due to changed APIs). Some pages like Syststrace are out of date/broken and will need to be opted out of typechecking (and given a warning) while they are in disrepair.

* Update scripts/lint-examples/bin/tsc-examples.js