feat: jumps to anchor in tabbed page (#3626)
We use tabs to nest whole pages. This means deep links might not aleady
have the anchor referenced id on the page. For example:

https://reactnative.dev/docs/environment-setup#ruby

Only applies if the 'React Native CLI Quickstart' -> macOS -> iOS tabs
have been selected. There's no guarantee this is the case until the
document has loaded and the application have updated the DOM.

This jump-to behaviour should only run on the the initial load.